# swagger.yaml
swagger: "2.0"
info:
  description: "This is a documentation for a API Rest using Quarkus"
  version: "1.0.0"
  title: "Quarkus Social"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "dougllasfps@email.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "quarkussocial.github.io"
basePath: "/"
...
paths:
...
  /users/{userId}/followers:
    put:
      tags:
        - "followers"
      summary: "Follow a user"
      parameters:
        - name: "userId"
          description: "the user identifier, as userId"
          in: path
          required: true
          type: "integer"

        - name: "Follower data"
          description: "the user identifier, as userId"
          in: body
          required: true
          schema:
            $ref: "#/definitions/FollowerRequest"
      responses:
        "200":
          description: "Ok"


    get:
      tags:
        - "followers"
      summary: "List followers of a user"
      parameters:
        - name: userId
          description: "the user identifier, as userId"
          in: path
          required: true
          type: "integer"

      responses:
        "200":
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/FollowersPerUserResponse"
    delete:
      tags:
        - "followers"
      summary: "Unfollow a user"
      parameters:
        - name: userId
          description: "the user identifier, as userId"
          in: path
          required: true
          type: "integer"
        - name: followerId
          description: "the follower identifier, as followerId"
          in: query
          required: true
          type: "integer"
      responses:
        "204":
          description: "Ok! No content"
definitions:
...
  FollowerRequest:
    type: "object"
    properties:
      followerId:
        type: "integer"
  FollowerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
  FollowersPerUserResponse:
    type: "object"
    properties:
      followersCount:
        type: "integer"
      content:
        type: "array"
        items:
          $ref: "#/definitions/FollowerResponse"